Solved Day 1 of Advent of Code
I have implemented solutions to the two puzzles from Day 1 of Advent of Code using the Crystal programming language. The first puzzle asks us to find the Elf carrying the most Calories, and the second puzzle asks us to find the total Calories carried by the top three Elves carrying the most Calories.

I have tested the solutions with the provided example input, and they produce the correct answers. I have also added a few additional test cases to ensure the solutions work for different input scenarios.

The solutions are implemented in two files called `day_1/puzzle_1.cr` and `day_1/puzzle_2.cr`, respectively. To compile the code, run the following command from the `day_1` directory:

$ crystal build puzzle_1.cr
$ crystal build puzzle_2.cr

Once the code is compiled, the solutions can be run from the command line by passing the input file as an argument, like this:

$ ./puzzle_1 real_input.txt
$ ./puzzle_2 real_input.txt

Note that in this commit message, I have renamed the input file from `input.txt` to `real_input.txt` to make it clear that this is the real input for the puzzle, and not just an example input. This is a common convention in Git, to differentiate between example inputs and real inputs.